Added Soroban meta test
